edit product

+++ src/app/admin/actions/Products.ts
@@ -24,6 +24,11 @@ const addSchema = z.object({
   }),
 });
 
+const editSchema = addSchema.extend({
+  file: fileSchema.optional(),
+  image: imageSchema.optional(),
+});
+
 export async function AddProduct(prevState: unknown, formData: FormData) {
   const result = addSchema.safeParse(Object.fromEntries(formData.entries()));
   if (!result.success) {
@@ -38,7 +43,10 @@ export async function AddProduct(prevState: unknown, formData: FormData) {
 
   await fs.mkdir("public/products", { recursive: true });
   const imagePath = `products/${crypto.randomUUID()}-${data.image.name}`;
-  await fs.writeFile(`public/${imagePath}`, Buffer.from(await data.image.arrayBuffer()));
+  await fs.writeFile(
+    `public/${imagePath}`,
+    Buffer.from(await data.image.arrayBuffer())
+  );
 
   await db.product.create({
     data: {
@@ -47,21 +55,67 @@ export async function AddProduct(prevState: unknown, formData: FormData) {
       description: data.description,
       priceInCents: data.priceInCents,
       filePath,
-      imagePath
+      imagePath,
     },
   });
 
   redirect("/admin/products");
 }
 
-export async function toggleProductAvailability(id: string, isAvailableForPurchase: boolean) {
+export async function updateProduct(
+  id: string,
+  prevState: unknown,
+  formData: FormData
+) {
+  const result = editSchema.safeParse(Object.fromEntries(formData.entries()));
+  if (!result.success) {
+    return result.error.formErrors.fieldErrors;
+  }
+
+  const data = result.data;
+  const product = await db.product.findUnique({ where: { id } });
+
+  if (product === null) return notFound();
+
+  let filePath = product.filePath;
+  if (data.file != null && data.file.size > 0) {
+    await fs.unlink(product.filePath);
+    filePath = `products/${crypto.randomUUID()}-${data.file.name}`;
+    await fs.writeFile(filePath, Buffer.from(await data.file.arrayBuffer()));
+  }
+
+  let imagePath = product.imagePath;
+  if (data.image != null && data.image.size > 0) {
+    await fs.unlink(`public/${product.imagePath}`);
+    imagePath = `products/${crypto.randomUUID()}-${data.image.name}`;
+    await fs.writeFile(`public/${imagePath}`, Buffer.from(await data.image.arrayBuffer()));
+  }
+
+  await db.product.update({
+    where: { id },
+    data: {
+      name: data.name,
+      description: data.description,
+      priceInCents: data.priceInCents,
+      filePath,
+      imagePath,
+    },
+  });
+
+  redirect("/admin/products");
+}
+
+export async function toggleProductAvailability(
+  id: string,
+  isAvailableForPurchase: boolean
+) {
   await db.product.update({ where: { id }, data: { isAvailableForPurchase } });
 }
 
 export async function deleteProduct(id: string) {
   const product = await db.product.delete({ where: { id } });
-  if(product === null) return notFound()
+  if (product === null) return notFound();
 
-  await fs.unlink(product.filePath)
-  await fs.unlink(`public/${product.imagePath}`)
+  await fs.unlink(product.filePath);
+  await fs.unlink(`public/${product.imagePath}`);
 }
